move index page load in a single loader+transaction
