+++ src/main/java/alfio/controller/IndexController.java
@@ -16,54 +16,36 @@
  */
 package alfio.controller;
 
-import alfio.config.authentication.support.OpenIdPrincipal;
-import alfio.controller.api.support.TicketHelper;
-import alfio.controller.api.v2.model.Language;
-import alfio.controller.api.v2.user.support.EventLoader;
-import alfio.controller.support.CSPConfigurer;
-import alfio.manager.PurchaseContextManager;
-import alfio.manager.i18n.MessageSourceManager;
-import alfio.manager.system.ConfigurationLevel;
-import alfio.manager.system.ConfigurationManager;
-import alfio.model.*;
-import alfio.model.system.ConfigurationKeys;
-import alfio.repository.*;
-import alfio.repository.user.OrganizationRepository;
-import alfio.util.ClockProvider;
-import alfio.util.Json;
-import alfio.util.MustacheCustomTag;
+import alfio.controller.support.DataPreloaderManager;
+import alfio.model.TicketReservationStatusAndValidation;
+import alfio.repository.EventRepository;
+import alfio.repository.SubscriptionRepository;
+import alfio.repository.TicketReservationRepository;
 import alfio.util.RequestUtils;
-import ch.digitalfondue.jfiveparse.*;
-import jakarta.servlet.http.HttpServletRequest;
+import ch.digitalfondue.jfiveparse.Document;
+import ch.digitalfondue.jfiveparse.JFiveParse;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import org.apache.commons.collections4.IterableUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.context.annotation.Profile;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
-import org.springframework.security.web.csrf.CsrfTokenRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.util.UriComponentsBuilder;
-import org.springframework.web.util.UriUtils;
 
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.nio.charset.StandardCharsets;
-import java.time.ZonedDateTime;
-import java.util.*;
-import java.util.regex.Pattern;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 import static alfio.config.Initializer.PROFILE_LIVE;
 import static alfio.controller.Constants.*;
-import static alfio.model.system.ConfigurationKeys.BASE_CUSTOM_CSS;
-import static alfio.util.HttpUtils.APPLICATION_JSON;
-import static java.util.Objects.requireNonNull;
 
 @Controller
 @Profile(PROFILE_LIVE)
@@ -71,49 +53,23 @@ public class IndexController {
 
     private static final String TEXT_HTML_CHARSET_UTF_8 = "text/html;charset=UTF-8";
     private static final String UTF_8 = "UTF-8";
-
     private final Document indexPage;
     private final Document openGraphPage;
-    private final ConfigurationManager configurationManager;
     private final EventRepository eventRepository;
-    private final FileUploadRepository fileUploadRepository;
-    private final MessageSourceManager messageSourceManager;
-    private final EventDescriptionRepository eventDescriptionRepository;
-    private final OrganizationRepository organizationRepository;
     private final TicketReservationRepository ticketReservationRepository;
     private final SubscriptionRepository subscriptionRepository;
-    private final EventLoader eventLoader;
-    private final PurchaseContextManager purchaseContextManager;
-    private final Json json;
-    private final CsrfTokenRepository csrfTokenRepository;
-    private final CSPConfigurer cspConfigurer;
+    private final DataPreloaderManager dataPreloaderManager;
 
-    public IndexController(ConfigurationManager configurationManager,
-                           EventRepository eventRepository,
-                           FileUploadRepository fileUploadRepository,
-                           MessageSourceManager messageSourceManager,
-                           EventDescriptionRepository eventDescriptionRepository,
-                           OrganizationRepository organizationRepository,
+    public IndexController(EventRepository eventRepository,
                            TicketReservationRepository ticketReservationRepository,
                            SubscriptionRepository subscriptionRepository,
-                           EventLoader eventLoader,
-                           PurchaseContextManager purchaseContextManager,
-                           CsrfTokenRepository csrfTokenRepository,
-                           CSPConfigurer cspConfigurer,
-                           Json json) {
-        this.configurationManager = configurationManager;
+                           DataPreloaderManager dataPreloaderManager) {
+
         this.eventRepository = eventRepository;
-        this.fileUploadRepository = fileUploadRepository;
-        this.messageSourceManager = messageSourceManager;
-        this.eventDescriptionRepository = eventDescriptionRepository;
-        this.organizationRepository = organizationRepository;
+
         this.ticketReservationRepository = ticketReservationRepository;
         this.subscriptionRepository = subscriptionRepository;
-        this.eventLoader = eventLoader;
-        this.purchaseContextManager = purchaseContextManager;
-        this.csrfTokenRepository = csrfTokenRepository;
-        this.cspConfigurer = cspConfigurer;
-        this.json = json;
+        this.dataPreloaderManager = dataPreloaderManager;
         try (var idxIs = new ClassPathResource("alfio-public-frontend-index.html").getInputStream();
              var idxOpenIs = new ClassPathResource("alfio/web-templates/event-open-graph-page.html").getInputStream();
              var idxIsR = new InputStreamReader(idxIs, StandardCharsets.UTF_8);
@@ -220,44 +176,21 @@ public class IndexController {
 
         response.setContentType(TEXT_HTML_CHARSET_UTF_8);
         response.setCharacterEncoding(UTF_8);
-        var nonce = cspConfigurer.addCspHeader(response, detectConfigurationLevel(eventShortName, subscriptionId), true);
 
-        if (eventShortName != null && RequestUtils.isSocialMediaShareUA(userAgent) && eventRepository.existsByShortName(eventShortName)) {
-            try (var os = response.getOutputStream(); var osw = new OutputStreamWriter(os, StandardCharsets.UTF_8)) {
-                var res = getOpenGraphPage((Document) openGraphPage.cloneNode(true), eventShortName, request, lang);
-                JFiveParse.serialize(res, osw);
-            }
-        } else {
-            try (var os = response.getOutputStream(); var osw = new OutputStreamWriter(os, StandardCharsets.UTF_8)) {
-                var baseCustomCss = configurationManager.getForSystem(BASE_CUSTOM_CSS).getValueOrNull();
-                var idx = indexPage.cloneNode(true);
-                if(authentication instanceof OAuth2AuthenticationToken oauth
-                    && oauth.getPrincipal() instanceof OpenIdPrincipal principal
-                    && principal.isSignedUp()
-                    && session.isNew()) {
-                    Optional.ofNullable(IterableUtils.get(idx.getElementsByTagName("html"), 0))
-                        .ifPresent(html -> html.setAttribute("data-signed-up", "true"));
-                }
-                idx.getElementsByTagName("script").forEach(element -> element.setAttribute(NONCE, nonce));
-                var head = idx.getElementsByTagName("head").get(0);
-                head.appendChild(buildScripTag(json.asJsonString(configurationManager.getInfo(session)), APPLICATION_JSON, "preload-info", null));
-                var httpServletRequest = requireNonNull(request.getNativeRequest(HttpServletRequest.class));
-                head.appendChild(buildMetaTag("GID", request.getSessionId()));
-                var csrf = csrfTokenRepository.loadToken(httpServletRequest);
-                if (csrf == null) {
-                    csrf = csrfTokenRepository.generateToken(httpServletRequest);
-                }
-                head.appendChild(buildMetaTag("XSRF_TOKEN", csrf.getToken()));
-                if (baseCustomCss != null) {
-                    var style = new Element("style");
-                    style.setAttribute("type", "text/css");
-                    style.appendChild(new Text(baseCustomCss));
-                    head.appendChild(style);
-                }
-                head.appendChild(buildMetaTag("authentication-enabled", Boolean.toString(configurationManager.isPublicOpenIdEnabled())));
-                preloadEventData(eventShortName, request, session, eventLoader, head, messageSourceManager, idx, json, lang);
-                JFiveParse.serialize(idx, osw);
-            }
+        try (var os = response.getOutputStream(); var osw = new OutputStreamWriter(os, StandardCharsets.UTF_8)) {
+            var doc = dataPreloaderManager.generateIndexDocument(
+                eventShortName,
+                subscriptionId,
+                userAgent,
+                lang,
+                request,
+                response,
+                session,
+                authentication,
+                openGraphPage,
+                indexPage
+            );
+            JFiveParse.serialize(doc, osw);
         }
     }
 
@@ -292,59 +225,6 @@ public class IndexController {
         }
     }
 
-    static void preloadEventData(String eventShortName,
-                                 ServletWebRequest request,
-                                 HttpSession session,
-                                 EventLoader eventLoader,
-                                 Element head,
-                                 MessageSourceManager messageSourceManager,
-                                 Node idx,
-                                 Json json,
-                                 String lang) {
-        String preloadLang = Objects.requireNonNullElse(lang, "en");
-        if (eventShortName != null) {
-            var eventInfoOptional = eventLoader.loadEventInfo(eventShortName, session);
-            if (eventInfoOptional.isPresent()) {
-                var ev = eventInfoOptional.get();
-                head.appendChild(buildScripTag(json.asJsonString(ev), APPLICATION_JSON, "preload-event", eventShortName));
-                preloadLang = getMatchingLocale(request, ev.getContentLanguages().stream().map(Language::getLocale).toList(), lang).getLanguage();
-                if (ZonedDateTime.now(ClockProvider.clock()).isAfter(((Event)ev.purchaseContext()).getEnd())) {
-                    // event is over.
-                    head.appendChild(buildMetaTag("robots", "noindex"));
-                }
-            }
-        }
-        head.appendChild(buildScripTag(json.asJsonString(messageSourceManager.getBundleAsMap("alfio.i18n.public", true, preloadLang, MessageSourceManager.PUBLIC_FRONTEND)), "application/json", "preload-bundle", preloadLang));
-        head.appendChild(buildScripTag(countriesForVatAsJson(json, preloadLang), "application/json", "preload-vat-countries", preloadLang));
-        head.appendChild(buildScripTag(countriesAsJson(json, preloadLang), "application/json", "preload-countries", preloadLang));
-        // add fallback in english
-        if (!"en".equals(preloadLang)) {
-            head.appendChild(buildScripTag(json.asJsonString(messageSourceManager.getBundleAsMap("alfio.i18n.public", true, "en", MessageSourceManager.PUBLIC_FRONTEND)), "application/json", "preload-bundle", "en"));
-        }
-        var htmlElement = IterableUtils.get(idx.getElementsByTagName("html"), 0);
-        htmlElement.setAttribute("lang", preloadLang);
-    }
-
-    private static String countriesForVatAsJson(Json json, String preloadLang) {
-        return json.asJsonString(TicketHelper.getSortedLocalizedVatCountries(Locale.forLanguageTag(preloadLang)));
-    }
-
-    private static String countriesAsJson(Json json, String preloadLang) {
-        return json.asJsonString(TicketHelper.getSortedLocalizedCountries(Locale.forLanguageTag(preloadLang)));
-    }
-
-    private static Element buildScripTag(String content, String type, String id, String param) {
-        var encodedContent = UriUtils.encodeFragment(content, StandardCharsets.UTF_8);
-        var e = new Element("script");
-        e.appendChild(new Text(encodedContent));
-        e.setAttribute("type", type);
-        e.setAttribute("id", id);
-        if (param != null) {
-            e.setAttribute("data-param", param);
-        }
-        return e;
-    }
-
     private static String reservationStatusToUrlMapping(TicketReservationStatusAndValidation status) {
         return switch (status.getStatus()) {
             case PENDING -> Boolean.TRUE.equals(status.getValidated()) ? "overview" : "book";
@@ -355,85 +235,6 @@ public class IndexController {
             case IN_PAYMENT, STUCK -> "error";
             default -> NOT_FOUND; // <- this may be a little bit aggressive
         };
-    }
-
-
-    /**
-     * Return the best matching locale.
-     *
-     * @param request
-     * @param contextLanguages list of languages configured for the event (o other contexts)
-     * @param lang override passed as parameter
-     * @return
-     */
-    private static Locale getMatchingLocale(ServletWebRequest request, List<String> contextLanguages, String lang) {
-        var locale = RequestUtils.getMatchingLocale(request, contextLanguages);
-        if (lang != null && contextLanguages.stream().anyMatch(lang::equalsIgnoreCase)) {
-            locale = Locale.forLanguageTag(lang);
-        }
-        return locale;
-    }
-
-    // see https://github.com/alfio-event/alf.io/issues/708
-    // use ngrok to test the preview
-    private Document getOpenGraphPage(Document eventOpenGraph, String eventShortName, ServletWebRequest request, String lang) {
-        var event = eventRepository.findByShortName(eventShortName);
-        var locale = getMatchingLocale(request, event.getContentLanguages().stream().map(ContentLanguage::getLanguage).toList(), lang);
-
-        var baseUrl = configurationManager.getForSystem(ConfigurationKeys.BASE_URL).getRequiredValue();
-
-        var title = messageSourceManager.getMessageSourceFor(event).getMessage("event.get-your-ticket-for", new String[] {event.getDisplayName()}, locale);
-
-        var head = eventOpenGraph.getElementsByTagName("head").get(0);
-
-        eventOpenGraph.getElementsByTagName("html").get(0).setAttribute("lang", locale.getLanguage());
-
-        //
-
-        getMetaElement(eventOpenGraph, "name", "twitter:image").setAttribute(CONTENT, baseUrl + "/file/" + event.getFileBlobId());
-        //
-
-        eventOpenGraph.getElementsByTagName("title").get(0).appendChild(new Text(title));
-        getMetaElement(eventOpenGraph, PROPERTY, "og:title").setAttribute(CONTENT, title);
-        getMetaElement(eventOpenGraph, PROPERTY,"og:image").setAttribute(CONTENT, baseUrl + "/file/" + event.getFileBlobId());
-
-        var eventDesc = eventDescriptionRepository.findDescriptionByEventIdTypeAndLocale(event.getId(), EventDescription.EventDescriptionType.DESCRIPTION, locale.toLanguageTag()).orElse("").trim();
-        var firstLine = Pattern.compile("\n").splitAsStream(MustacheCustomTag.renderToTextCommonmark(eventDesc)).findFirst().orElse("");
-        getMetaElement(eventOpenGraph, PROPERTY,"og:description").setAttribute(CONTENT, firstLine);
-
-
-        var org = organizationRepository.getById(event.getOrganizationId());
-        var author = String.format("%s <%s>", org.getName(), org.getEmail());
-        getMetaElement(eventOpenGraph, "name", "author").setAttribute(CONTENT, author);
-
-        fileUploadRepository.findById(event.getFileBlobId()).ifPresent(metadata -> {
-            var attributes = metadata.getAttributes();
-            if (attributes.containsKey(FileBlobMetadata.ATTR_IMG_HEIGHT) && attributes.containsKey(FileBlobMetadata.ATTR_IMG_WIDTH)) {
-                head.appendChild(buildOGMetaTag("og:image:width", attributes.get(FileBlobMetadata.ATTR_IMG_WIDTH)));
-                head.appendChild(buildOGMetaTag("og:image:height", attributes.get(FileBlobMetadata.ATTR_IMG_HEIGHT)));
-            }
-        });
-
-        return eventOpenGraph;
-    }
-
-    private static Element buildOGMetaTag(String propertyValue, String contentValue) {
-        return buildMetaTag(PROPERTY, propertyValue, contentValue);
-    }
-
-    private static Element buildMetaTag(String name, String content) {
-        return buildMetaTag("name", name, content);
-    }
-
-    private static Element buildMetaTag(String property, String propertyValue, String content) {
-        var meta = new Element("meta");
-        meta.setAttribute(property, propertyValue);
-        meta.setAttribute(CONTENT, content);
-        return meta;
-    }
-
-    private static Element getMetaElement(Document document, String attrName, String propertyValue) {
-        return (Element) document.getAllNodesMatching(Selector.select().element("meta").attrValEq(attrName, propertyValue).toMatcher(), true).get(0);
     }
 
     @GetMapping(value = {
@@ -456,8 +257,4 @@ public class IndexController {
         return REDIRECT + UriComponentsBuilder.fromPath("/event/{eventShortName}").build(Map.of(EVENT_SHORT_NAME, eventName));
     }
 
-    private ConfigurationLevel detectConfigurationLevel(String eventShortName, String subscriptionId) {
-        return purchaseContextManager.detectConfigurationLevel(eventShortName, subscriptionId)
-            .orElseGet(ConfigurationLevel::system);
-    }
 }
